fixed camera power on/ off bug added opcode

+++ src/Monitors/CameraReportMonitor.hpp
@@ -10,7 +10,7 @@ public:
     CameraReportMonitor(unsigned int offset);
     void execute();
     File imgFile;
-    void create_camera_report(int fragment_number, short serial_number);
+    void create_camera_report(int fragment_number, uint8_t serial_number);
     void add_possible_command();
 };
 #endif
